requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
aiofiles==23.1.0
python-multipart==0.0.6
PyPDF2==3.0.0
pdf2image==1.16.0
pillow==10.0.0
reportlab==4.0